build/make: support whitespace in file names
The mirroring of the source tree via symlinks failed to handle file
names with whitespaces correctly. This can be observed with the git
tarball releases. The problem was caused by being to sloppy with parsing
the output of the find command. It is fixed by interpreting each line as
one file name.