import: prevent accidental deletion of {src/,raw/}
On import, existing src/ and raw/ folders were deleted irrespective of
whether they are populated by the import. In consequence, raw/ could not
be used in conjunction with import.

With this fix, src/ and raw/ are only deleted if the contrib directory
contains a src/ or raw/ subdirectory respectively.

Fixes nfeske/goa#16