Support for AARCH64
This patch adds support for 64-bit ARM in addition to x86, and the
ability to cross compile between architectures.

By default, Goa targets the architecture of the host system, currently
supporting x86_64 and aarch64. Alternatively, the architecture can
be specified via the new '--arch' command-line argument. Supported
values are 'x86_64' and 'arm_v8a', which correspond to naming used in
Genode's depot. The values can also be assigned to the 'arch'
configuration variable in a .goarc file.

Fixes #18