Refine libgcc linker handling
This patch replaces the former use of -nostdlib by the combination of
-nostartfiles, -nolibc, and -static-libgcc. This has two benefits.

First, we don't need to query the path of libgcc for specifying libgcc.a
at the linker command line anymore. GCC knowns where to find its libgcc
after all.

Second, GCCs builtin heuristics for the linkage of libgcc comes into
effect.  In particular, in some cases, libgcc needs to be specified
after static libraries when linking a target - if one of the static
libraries depends on libgcc symbols. This, however, is not always
possible. For example, when using CMake, ldlibs and ldflags may appear
before the static libraries. When using GCCs builtin heuristics for
linking libgcc, we side-step such limitations.

For reference, the specific issue was the linking of Hatari on AARCH64.
Here, the linker reported undefined references to libgcc's __addtf3 (and
others) for cpu/libUaeCpu.a (fpp_native.c, fdc.c). The symbols would not
be resolved because libgcc.a appeared too early at the linker command
line.

Issue #18