add Vivado support
This commit adds support for recreating Xilinx Vivado projects and generating
bitstreams. For this purpose, Goa detects the presence of a vivado.tcl
file in the src/ directory. This file must be created in Vivado with the
following command:

write_project_tcl -paths_relative_to /path/to/workspace/dir vivado.tcl

Furthermore, the required source files (as referenced in the
checkRequiredFiles procedure in the tcl file) must be manually copied into
the src/ directory.

Note that the Vivado tools must have been installed and sourced.