linux.tcl: add nic_router for Nic requirements
The linux_nic_drv previously had a nic_server-mode that was used to
implement the NIC requirements. Since removal of this mode from the
driver, the support for NIC sessions is broken in Goa. As the NIC driver
solely operates as an Uplink client, we need to additionally instantiate
a NIC router that we use as a NIC server for the runtime.

genodelabs/goa#38