support index projects
An index projects is defined by an `index` file. This file follows the
structure of a depot index but only names the archive names (lacking
depot user and version). The `export` command augments these names with
the current depot user and version information. The result is published
as a depot index for the current Sculpt version.

genodelabs/goa#39