run/linux: add nitpicker background domain
For testing transparency of GUI applications in Goa, it is useful to add a
background domain to nitpicker's config. Analogously to Sculpt, GUI sessions
with the label "backdrop" will be assigned to this domain.

Special care needs to be taken for clients using Command::Backdrop
because their view will not be shown unless there is another view owning
the focus in the same domain. Rewriting the session label to "backdrop",
however, instructs nitpicker to use the view as the default background.

genodelabs/goa#70