improve hint about possibly missing artifacts file
The hint was not displayed when starting with a bare project and an
_import_ file. Running `goa build` in this case would create a src/
directory after the check for a missing _artifacts_ file.

Instead of not considering a directory as a goa project when src/ but no
_artifacts_ file is present, we still consider it when an _import_ file
is present. Note that disregarding a directory that merely contains a
src/ directory without neither of these two file is required since "src"
is a very common directory name.

Changing the looks_like_goa_project logic allows moving the check for a
missing _artifacts_ file to a after the import was done.

genodelabs/goa#49