export: correctly pass arch and pkg name
When exporting an index project, we must ensure that arch is passed correctly
for the recursive export. Otherwise, Goa might fail to export pkg archives for
projects that have already been exported for other architectures.
Moreover, we must pass the pkg name since a project might contain other pkg
runtimes that we do not need.

Fixes #90