index: prefer goarc-defined versions
On index export/publish, the goarc-defined versions should be preferred
over the versions defined in a Goa-project's 'version' file. This allows
pinning pkg archives in an index to a particular version. This change
also requires that it is checked wether the expected version will be exported
when performing a recursive export.