autoconf: shared lib flag support
For autoconf, there is no general way to set particular flags for
building shared libraries. With this commit, we propagate the flags to
'./configure' and 'make' in form of the environment variable LDLIBS_SO.
In most cases, the configure-script requires patching to make use of
this variable, though. A good entry point for patching is to look for the
'archive_cmds' variable that is passed to libtool.