export: fix import of dependent archives
If both, a src and bin archive, are missing, the depot/dependencies tool
will only output the src archive path. We therefore need to make sure to
adapt the path and download the corresponding bin archive (along with the src
archive). Otherwise, the user is faced with an unexplainable download of
bin archives on a second 'goa export'.