update-goa: disable sandbox when fetching via SSH
If git's origin remote url is not https, we assume it is SSH and ask
the user for confirmation to disable sandboxing. This is required because
getting SSH to work reliably in bwrap is quite complicated. Moreover,
we do not want to add a mechanism that makes the user's keypair available
in the sandbox. This seems overkill just for Goa developers to do git
fetch in a sandboxed environment.

Issue #136