query version information from depot-index files
When supplied with the `--update-index` command-line argument or when
the `update_config` variable is set to 1, Goa automatically downloads
depot index files in order to look-up archive versions. This is
only done as a last resort, i.e. if there is no version information
defined in _goarc_ files and if Goa did not find a corresponding project
directory.

Issue #117